refactor(tui): extract shared format utils and TypeBadge component
Deduplicate formatRelativeTime, formatTimestamp, TypeBadge, and
PriorityIndicator into shared modules. Add selectedIndex reset when
message list or filter changes. Add explanatory comment to listAll
default limit.
