+++ apps/tui/src/utils/format.ts
@@ -1,0 +1,18 @@
+export function formatRelativeTime(timestamp: number | null): string {
+  if (!timestamp) return 'never';
+  const diff = Date.now() - timestamp;
+  const seconds = Math.floor(diff / 1000);
+  if (seconds < 60) return `${seconds}s ago`;
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  if (days > 0) return `${days}d ago`;
+  return 'offline';
+}
+
+export function formatTimestamp(timestamp: number): string {
+  const date = new Date(timestamp);
+  return date.toLocaleString();
+}
